Resolved for circular dependency
