+++ api/db/dbController.js
@@ -52,31 +52,6 @@ const Rollback = async () => {
 };
 
 /**
- * Function to generate create query
- * @param {string} query
- * @param {object} req
- * @returns {string}
- */
-const GenerateCreateQuery = async (query, req) => {
-    const createdBy = req?.authorizedUser?.userId || 0;
-    const updatedQuery = `${query.replace("{{CREATED_BY}}", createdBy)} RETURNING *`;
-    return updatedQuery
-};
-
-/**
- * Function to generate modify query
- * @param {string} query
- * @param {object} req
- * @returns {string}
- */
-const GenerateModifyQuery = async (query, req) => {
-    const modifiedOn = getCurrentDateTime();
-    const modifiedBy = req?.authorizedUser?.userId || 0;
-    const updatedQuery = `${query}, "ModifiedOn" = $${modifiedOn}, "ModifiedBy" = $${modifiedBy} RETURNING *`;
-    return updatedQuery;
-};
-
-/**
  * Function to get user by username
  * @param {string} username
  * @returns {Promise}
@@ -412,8 +387,6 @@ module.exports = {
     Begin,
     Commit,
     Rollback,
-    GenerateCreateQuery,
-    GenerateModifyQuery,
     GetUserByUsername,
     UpdateResetCode,
     UpdatePasswordAgainstUsername,
